learning_langchain/pyproject.toml
[project]
name = "learning_langchain"
version = "0.1.0"
description = "A hands-on project exploring LangChain's integration with LLM providers like OpenAI, Anthropic, Hugging Face, and Google Gemini. Designed to build, test, and experiment with LLM applications using the latest LangChain 0.3+ ecosystem."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "grandalf>=0.8",
    "huggingface-hub>=0.34.3",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.18",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.72",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.28",
    "numpy>=2.3.2",
    "openai>=1.99.0",
    "pydantic>=2.11.7",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "streamlit>=1.48.0",
    "torch>=2.9.1",
    "transformers>=4.55.0",
]
